Add files via upload
This commit should hopefully give a system of file structure, aka, tell people where to plug their files in to be used in development of the game. I've also included README files for the folders for the sake of guidance of structure.

All of this is entirely open to advice / recommendations / requests / etc.